+++ packages/blade-svelte/src/components/index.ts
@@ -23,8 +23,13 @@ export { default as Spinner } from './Spinner/BaseSpinner/BaseSpinner.svelte';
 // Amount
 export { default as Amount } from './Amount/Amount.svelte';
 
+// Icons
+export * from './Icons';
 // Badge
 export { default as Badge } from './Badge/Badge.svelte';
+
+// Divider
+export { default as Divider } from './Divider/Divider.svelte';
 
 // BladeProvider
 export * from './BladeProvider';